后台文章模块

+++ weblog-web/src/main/java/com/szq/web/controller/admin/AdminCategoryController.java
@@ -1,11 +1,15 @@
 package com.szq.web.controller.admin;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.szq.web.mapper.ArticleCategoryRelMapper;
 import com.szq.web.mapper.CategoryMapper;
+import com.szq.web.model.ArticleCategoryRel;
 import com.szq.web.service.admin.AdminCategoryService;
 import com.szq.web.model.Category;
 import com.szq.web.utils.BaseResponse;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -14,14 +18,15 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/admin/category")
+@RequiredArgsConstructor
 public class AdminCategoryController {
 
-    @Autowired
-    private AdminCategoryService categoryService;
-    @Autowired
-    private CategoryMapper categoryMapper;
-    @Autowired
-    private List<ObjectMapper> objectMappers;
+    private final AdminCategoryService categoryService;
+
+    private final CategoryMapper categoryMapper;
+
+    private final ArticleCategoryRelMapper articleCategoryRelMapper;
+
 
     /**
      * 添加分类
@@ -55,6 +60,12 @@ public class AdminCategoryController {
         if (category1 == null){
             return BaseResponse.fail("分类不存在");
         }
+        Long count = articleCategoryRelMapper.selectCount(new LambdaQueryWrapper<>(ArticleCategoryRel.class)
+                .eq(ArticleCategoryRel::getCategoryId, id));
+        if (count > 0){
+            return BaseResponse.fail("分类还有关联不可以删除");
+        }
+
         return categoryService.removeById(id) ? BaseResponse.success() : BaseResponse.fail();
     }
 
@@ -62,7 +73,7 @@ public class AdminCategoryController {
     /**
      * 查询所有分类
      */
-    @GetMapping("/select/list")
+    @PostMapping("/select/list")
     public BaseResponse<Object> selectList() {
         List<Category> list = categoryMapper.selectList(null);
         return BaseResponse.success(list);
